refactor dummy data
